refactor: minor fixes & improvements
