+++ package.json
@@ -21,6 +21,7 @@
     "cookie-session": "^1.4.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "lodash": "^4.17.20",
     "mongoose": "^5.10.6",
     "passport": "^0.4.1",
     "passport-facebook": "^3.0.0",
