+++ server/middleware/validation.js
@@ -1,9 +1,9 @@
-import * as yup from 'yup';
+import { validateInput } from '../services/yup';
 
 export const validate = schema => {
   return async (req, res, next) => {
     try {
-      const body = await yup.object().shape(schema).strict(true).noUnknown().validate(req.body, { abortEarly: false });
+      const body = await validateInput(schema, req.body);
       req.body = body;
       next();
     } catch (err) {
