+++ server/models/index.js
@@ -1,0 +1,2 @@
+export * from './List';
+export * from './User';
