+++ server/routes/index.js
@@ -1,0 +1,2 @@
+export { default as authRouter } from './authRoutes';
+export { default as listRouter } from './listRoutes';
