+++ server/services/yup.js
@@ -1,5 +1,9 @@
 import * as yup from 'yup';
 import { StatusEnum, TypeEnum } from '../constants';
+import { IsValidMongoId } from '../utils';
+
+export const validateInput = (schema, validationObj) =>
+  yup.object().shape(schema).strict(true).noUnknown().validate(validationObj, { abortEarly: false });
 
 export const signupSchema = {
   name: yup.string().min(3).max(50).trim().required(),
@@ -27,4 +31,8 @@ export const recordSchema = {
   year: yup.string(),
   imdbRating: yup.mixed().test('is-decimal', 'not a decimal number', val => !isNaN(val)),
   poster: yup.string()
+};
+
+export const idSchema = {
+  id: yup.string().test('validateMongoId', 'Invalid MongoId', IsValidMongoId).required()
 };
